+++ tests/testthat/test-linalg.R
@@ -154,7 +154,7 @@ test_that("solve", {
   A <- torch_randn(3, 3)
   b <- torch_randn(3)
   x <- linalg_solve(A, b)
-  expect_equal_to_tensor(torch_matmul(A, x), b, tolerance = 1e-6)
+  expect_equal_to_tensor(torch_matmul(A, x), b, tolerance = 1e-5)
 })
 
 test_that("lstsq", {