+++ NEWS.md
@@ -17,6 +17,7 @@
 - `torch_manual_seed()` now matches PyTorch's behavior so we can more easily compare implementations. Since this is a breaking change we added the `torch.old_seed_behavior=TRUE` option so users can stick to the old behavior. (#639)
 - Implemented `traced_module$graph_for()` to allow inspecting the optimized jit graph. (#643)
 - Improved CUDA version auto-detection (#644)
+- Added `slc` to allow dynamically creating slices when indexing tensors. (#648)
 
 # torch 0.4.0